fix: Resolve parameter naming inconsistencies and runtime errors
Phase 2 fixes:
- Added observe() method to BaseObservation (environment/base.py)
- Fixed dtype mismatch: float16->float32 in env_wrapper._to_tensor()
- Fixed pbar attribute error in Experiment.__del__()
- Fixed RBFDynamicsEnsemble/MLPDynamicsEnsemble to pass state_dim in dynamics_config
- Fixed test parameter naming in test_core_agent.py, test_core_experiment.py, test_policy.py
- Added IdentityActionEncoder to test_policy.py for MPC tests

Test results: 166 passed, 162 failing (up from 150 passed initially)

+++ actdyn/environment/env_wrapper.py
@@ -48,7 +48,7 @@ class EnvWrapper(gym.Wrapper):
         # Update observation space if needed
         if hasattr(obs_model, "d_obs"):
             self.observation_space = spaces.Box(
-                low=-np.inf, high=np.inf, shape=(obs_model.d_obs,), dtype=np.float16
+                low=-np.inf, high=np.inf, shape=(obs_model.d_obs,), dtype=np.float32
             )
 
     def _is_torch_native_env(self) -> bool:
@@ -63,9 +63,9 @@ class EnvWrapper(gym.Wrapper):
         """Convert input to torch tensor."""
         # Get the device from the observation model's network if available
         if isinstance(x, torch.Tensor):
-            x = x.to(self.device)
+            x = x.to(self.device).float()
         else:
-            x = torch.tensor(x, device=self.device, dtype=torch.float16)
+            x = torch.tensor(x, device=self.device, dtype=torch.float32)
 
         if x.dim() == 1:
             x = x.unsqueeze(0).unsqueeze(0)  # Add batch dimension if needed
